Update score type #1

+++ server/src/main/kotlin/com/coolcode/server/controller/EvaluationController.kt
@@ -7,7 +7,6 @@ import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import java.math.BigDecimal
 
 @RestController
 @RequestMapping("/api/evaluation")
@@ -15,7 +14,7 @@ class EvaluationController(
     private val evaluationService: EvaluationService,
 ) {
     @GetMapping("/username/{username}")
-    fun evaluateByUsername(@PathVariable("username") username: String): ResponseEntity<BigDecimal> {
+    fun evaluateByUsername(@PathVariable("username") username: String): ResponseEntity<Int> {
         return ResponseEntity(evaluationService.evaluate(username), HttpStatus.OK)
     }
 }
